MacOS - Fix incorrect File usage & improve docs.rs detection & Fix IPv6 (#19)
* don't use a file + better docs.rs detection

* MacOS: Fix IPv6 routes

An route message is a rt_msdhdr followed by a list of `struct
sockaddr`. However, the sockaddr can be variable length (in particular
because sockaddr_in6 is larger than sockaddr/sockaddr_in). This commit
fixes the logic of walking the list of sockaddr's.
Also did some refactoring to reduce the scope of some `unsafe` blocks.

Many of these chagnes are based on @adamierymenko's
https://github.com/johnyburd/net-route/pull/4. However that PR had some
bugs of its own.

The logic I'm using is inspired by
https://opensource.apple.com/source/network_cmds/network_cmds-606.40.2/netstat.tproj/route.c.auto.html

* MacOS: handle broken IPv6 link-local gateways

Apparently the kernel does something hacky and encodes the v6 scope_id of
a link_local gateway IP in bytes 2 & 3. This makes these IPs invalid
IPv6 addresses. Fix it by setting these bytes to 0.

* fix silly bug

* fix ipv6 fix asserts

* comments

---------

Co-authored-by: Gregor Maier <gregor@majordomus.org>